refactor(popup): replaced rux-pop-up with beta-pop-up, renamed beta
