--- packages/web-components/src/components/rux-beta-menu-item/rux-beta-menu-item.scss
@@ -1,18 +1,0 @@
-:host {
-    display: block;
-}
-
-.rux-menu-item {
-    position: relative;
-    padding: 6px 8px;
-    color: var(--color-interactive-default);
-    cursor: pointer;
-    &:hover {
-        color: var(--color-interactive-hover);
-        background: var(--color-background-base-hover);
-    }
-    &--selected {
-        background: var(--color-background-surface-selected);
-        color: var(--color-text-primary);
-    }
-}
