--- packages/web-components/src/components/rux-beta-menu-item/rux-beta-menu-item.tsx
@@ -1,25 +1,0 @@
-import { Prop, Component, Host, h } from '@stencil/core'
-
-@Component({
-    tag: 'rux-beta-menu-item',
-    styleUrl: 'rux-beta-menu-item.scss',
-    shadow: true,
-})
-export class RuxBetaMenuItem {
-    @Prop() selected = false
-
-    render() {
-        return (
-            <Host>
-                <div
-                    class={{
-                        'rux-menu-item': true,
-                        'rux-menu-item--selected': this.selected,
-                    }}
-                >
-                    <slot></slot>
-                </div>
-            </Host>
-        )
-    }
-}
