--- packages/web-components/src/components/rux-beta-pop-up-menu/rux-beta-pop-up-menu.scss
@@ -1,55 +1,0 @@
-:host {
-    display: block;
-}
-
-.hidden,
-:host([hidden]) {
-    display: none;
-}
-
-.rux-popup__content {
-    position: absolute;
-    color: white;
-    padding: 12px;
-    display: none;
-    border-radius: var(--radius-base);
-    background: var(--color-background-base-default);
-    border: 1px solid var(--color-interactive-muted);
-}
-
-.rux-popup__trigger {
-    display: contents;
-}
-
-.rux-popup__content--menu {
-    padding: 0;
-}
-.rux-popup__arrow-left {
-    border-right: 2px solid var(--color-interactive-muted);
-}
-
-.rux-popup__arrow-top {
-    border-bottom: 2px solid var(--color-interactive-muted);
-}
-.rux-popup__arrow-right {
-    border-left: 2px solid var(--color-interactive-muted);
-}
-.rux-popup__arrow-bottom {
-    border-top: 2px solid var(--color-interactive-muted);
-}
-
-.rux-popup {
-    // position: relative;
-}
-
-.rux-popup-arrow {
-    position: absolute;
-    background: #2b659b;
-    // background: linear-gradient(315deg, #232629 50%, #101923 50%); works great for a top placed menu, not for others lol
-    // width: 9px;
-    // height: 9px;
-    width: 17px;
-    height: 17px;
-    transform: rotate(45deg);
-    z-index: -9999;
-}
