--- packages/web-components/src/components/rux-beta-pop-up-menu/test/index.html
@@ -1,88 +1,0 @@
-<!DOCTYPE html>
-<html dir="ltr" lang="en">
-    <head>
-        <meta charset="utf-8" />
-        <meta
-            name="viewport"
-            content="width=device-width, initial-scale=1.0, minimum-scale=1.0, maximum-scale=5.0"
-        />
-        <title>Stencil Component Starter</title>
-
-        <link rel="preconnect" href="https://fonts.gstatic.com" />
-        <link
-            href="https://fonts.googleapis.com/css2?family=Roboto+Mono:wght@100;300;400&family=Roboto:wght@100;300;400;500;700;900&display=swap"
-            rel="stylesheet"
-        />
-        <script type="module" src="/build/astro-web-components.esm.js"></script>
-        <script nomodule src="/build/astro-web-components.js"></script>
-
-        <link rel="stylesheet" href="/build/astro-web-components.css" />
-    </head>
-
-    <body>
-        <div style="padding: 10% 5%">
-            <rux-beta-pop-up-menu placement="top-start">
-                <rux-button slot="trigger">Top</rux-button>
-                <rux-beta-menu id="test">
-                    <rux-beta-menu-item selected
-                        >Pop up menu option test</rux-beta-menu-item
-                    >
-                    <rux-beta-menu-item
-                        >Pop up menu option test</rux-beta-menu-item
-                    >
-                    <rux-beta-menu-item
-                        >Pop up menu option test</rux-beta-menu-item
-                    >
-                </rux-beta-menu>
-            </rux-beta-pop-up-menu>
-        </div>
-        <div style="padding: 5%">
-            <rux-beta-pop-up-menu placement="right-start">
-                <rux-button slot="trigger">Right</rux-button>
-                <rux-beta-menu id="test">
-                    <rux-beta-menu-item selected
-                        >Pop up menu option test</rux-beta-menu-item
-                    >
-                    <rux-beta-menu-item
-                        >Pop up menu option test</rux-beta-menu-item
-                    >
-                    <rux-beta-menu-item
-                        >Pop up menu option test</rux-beta-menu-item
-                    >
-                </rux-beta-menu>
-            </rux-beta-pop-up-menu>
-        </div>
-        <div style="padding: 5%">
-            <rux-beta-pop-up-menu placement="bottom-start">
-                <rux-button slot="trigger">Bottom</rux-button>
-                <rux-beta-menu id="test">
-                    <rux-beta-menu-item selected
-                        >Pop up menu option test</rux-beta-menu-item
-                    >
-                    <rux-beta-menu-item
-                        >Pop up menu option test</rux-beta-menu-item
-                    >
-                    <rux-beta-menu-item
-                        >Pop up menu option test</rux-beta-menu-item
-                    >
-                </rux-beta-menu>
-            </rux-beta-pop-up-menu>
-        </div>
-        <div style="padding: 5% 20%">
-            <rux-beta-pop-up-menu placement="left-start">
-                <rux-button slot="trigger">Left</rux-button>
-                <rux-beta-menu id="test">
-                    <rux-beta-menu-item selected
-                        >Pop up menu option test</rux-beta-menu-item
-                    >
-                    <rux-beta-menu-item
-                        >Pop up menu option test</rux-beta-menu-item
-                    >
-                    <rux-beta-menu-item
-                        >Pop up menu option test</rux-beta-menu-item
-                    >
-                </rux-beta-menu>
-            </rux-beta-pop-up-menu>
-        </div>
-    </body>
-</html>
