+++ packages/web-components/src/components/rux-menu-item/rux-menu-item.scss
@@ -1,78 +1,18 @@
 :host {
-    /**
-      * @prop --popup-menu-text-color: Menu Text Color
-    */
-    --popup-menu-text-color: var(--color-interactive-default);
-
-    /**
-      * @prop --popup-menu-item-background-color: Menu Item Background Color
-    */
-    --popup-menu-item-background-color: var(--color-background-base-default);
-
-    /**
-      * @prop --popup-menu-item-hover-background-color: Menu Item Hover Background Color
-    */
-    --popup-menu-item-hover-background-color: var(
-        --color-background-surface-hover
-    );
-
-    /**
-      * @prop --popup-menu-item-hover-background-color: Menu Item Hover Text Color
-    */
-    --popup-menu-item-hover-text-color: var(--color-interactive-hover);
-
     display: block;
+}
 
-    li {
-        list-style: none;
-        padding: 0.15rem 0.75rem;
-        overflow: hidden;
-        font-family: var(--font-body-1-font-family);
-        font-size: var(--font-body-1-font-size);
-        font-weight: var(--font-body-1-font-weight);
-        letter-spacing: var(--font-body-1-letter-spacing);
-        background-color: var(--popup-menu-item-background-color);
-        color: var(--popup-menu-text-color);
-        &:hover {
-            background-color: var(--popup-menu-item-hover-background-color);
-        }
-
-        a,
-        div {
-            text-overflow: ellipsis;
-            text-decoration: none;
-            color: var(--popup-menu-text-color);
-            word-wrap: none;
-            white-space: nowrap;
-            overflow: hidden;
-            &:hover {
-                color: var(--popup-menu-item-hover-text-color);
-            }
-        }
-
-        a {
-            display: block;
-            padding: 0;
-            width: 100%;
-        }
+.rux-menu-item {
+    position: relative;
+    padding: 6px 8px;
+    color: var(--color-interactive-default);
+    cursor: pointer;
+    &:hover {
+        color: var(--color-interactive-hover);
+        background: var(--color-background-base-hover);
     }
-}
-
-:host([hidden]) {
-    display: none;
-}
-
-:host([disabled]) {
-    cursor: not-allowed;
-
-    li {
-        &:hover {
-            background-color: transparent;
-        }
-
-        a,
-        div {
-            opacity: var(--disabled-opacity);
-        }
+    &--selected {
+        background: var(--color-background-surface-selected);
+        color: var(--color-text-primary);
     }
 }
