+++ packages/web-components/src/index.html
@@ -19,5 +19,24 @@
         <link rel="stylesheet" href="/build/astro-web-components.css" />
     </head>
 
-    <body></body>
+    <body>
+        <div style="margin: 15%">
+            <rux-pop-up-menu id="test" placement="top-start">
+                <rux-button slot="trigger">top</rux-button>
+                <rux-menu>
+                    <rux-menu-item selected
+                        >Pop up menu option test adsfasdf asdf asdf asdf asdf
+                    </rux-menu-item>
+                    <rux-menu-item
+                        >Pop up menu option testasdf asdf asdf asdf asdf asdf
+                        asdf
+                    </rux-menu-item>
+                    <rux-menu-item
+                        >Pop up menu option test asdf asdf asdf asdf asdf asdf
+                        asdf asdf
+                    </rux-menu-item>
+                </rux-menu>
+            </rux-pop-up-menu>
+        </div>
+    </body>
 </html>
